Expose database repr, database fixup
Parts of the database representation are exposed.
A fixup command for builder-migrations is added to remove bad database
entries fixed in a57798f4c02eb4d528b90932ec26fb0b718f1a13. It is up to
the operator to remove the files and optionally re-upload the 'full'
files to builder-web.