Add more json to endpoints, populate-local.sh
Some basic json is added to the endpoints that list jobs and builds.

With this a script populate-local.sh can be used to fetch a number of
builds from a remote instance and upload them to a local instance. This
is useful for populating an instance with real builds for testing
purposes.

Closes #11.