Work around dream-encoding footgun
Dream-encoding adds a Transfer-Encoding header to a response. If that
response is an error code it will go through the error handler which may
change the response - but the new response *does not* go through the
dream-encoding middleware! This means we may rewrite the response with a
plaintext message but still have the Transfer-Encoding header! This
makes clients very unhappy.